gentb_website/requirements/base.txt
Django==1.8
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
requests==2.4.1
pytz==2015.4
